+++ packages/open-next/src/adapters/util.ts
@@ -1,13 +1,62 @@
 import fs from "node:fs";
 import path from "node:path";
+import type { NextConfig, RoutesManifest } from "./next-types.js";
+import type { PublicFiles } from "../build.js";
 
 export function setNodeEnv() {
   process.env.NODE_ENV = process.env.NODE_ENV ?? "production";
+}
+
+export function generateUniqueId() {
+  return Math.random().toString(36).slice(2, 8);
 }
 
 export function loadConfig(nextDir: string) {
   const filePath = path.join(nextDir, "required-server-files.json");
   const json = fs.readFileSync(filePath, "utf-8");
   const { config } = JSON.parse(json);
-  return config;
+  return config as NextConfig;
+}
+
+export function loadHtmlPages(nextDir: string) {
+  const filePath = path.join(nextDir, "server", "pages-manifest.json");
+  const json = fs.readFileSync(filePath, "utf-8");
+  return Object.entries(JSON.parse(json))
+    .filter(([_, value]) => (value as string).endsWith(".html"))
+    .map(([key]) => key);
+}
+
+export function loadPublicAssets(openNextDir: string) {
+  const filePath = path.join(openNextDir, "public-files.json");
+  const json = fs.readFileSync(filePath, "utf-8");
+  return JSON.parse(json) as PublicFiles;
+}
+
+export function loadRoutesManifest(nextDir: string) {
+  const filePath = path.join(nextDir, "routes-manifest.json");
+  const json = fs.readFileSync(filePath, "utf-8");
+  const routesManifest = JSON.parse(json) as RoutesManifest;
+  // Static routes take precedence over dynamic routes
+  return [...routesManifest.staticRoutes, ...routesManifest.dynamicRoutes];
+}
+
+export function loadAppPathsManifestKeys(nextDir: string) {
+  const appPathsManifestPath = path.join(
+    nextDir,
+    "server",
+    "app-paths-manifest.json"
+  );
+  const appPathsManifestJson = fs.existsSync(appPathsManifestPath)
+    ? fs.readFileSync(appPathsManifestPath, "utf-8")
+    : "{}";
+  const appPathsManifest = JSON.parse(appPathsManifestJson) as Record<
+    string,
+    string
+  >;
+  return Object.keys(appPathsManifest).map((key) => {
+    // Remove group route params and /page suffix
+    const cleanedKey = key.replace(/\/\(\w+\)|\/page$/g, "");
+    // We need to check if the cleaned key is empty because it means it's the root path
+    return cleanedKey === "" ? "/" : cleanedKey;
+  });
 }